pretty the layout of prediction page
